[Feat/#31] MyPageCell - 내 포인트 레이아웃 설정
